added weekly forecast, updated current forecast ui

+++ app/src/main/java/com/example/weatherapp/model/Weather.kt
@@ -3,7 +3,7 @@ package com.example.weatherapp.model
 import com.google.gson.annotations.SerializedName
 
 data class Weather(
-    @SerializedName("id") val id:Float,
+    @SerializedName("id") val id:Int,
     @SerializedName("main") val main:String,
     @SerializedName("description") val description:String,
     @SerializedName("icon") val icon:String,
